+++ {{cookiecutter.project_slug}}/src/{{cookiecutter.project_slug}}.Application/Services/SubDomainAService.cs
@@ -1,0 +1,87 @@
+﻿// If necessary incorporate dependencies from the infrastructure most likely and persistence layers and adjust 
+// {{cookiecutter.project_slug}}.Application.csproj accordingly, e.g. in case if the persistence layer and ORMs are not required.
+
+using {{cookiecutter.project_slug}}.Domain.SubDomainA;
+
+namespace {{cookiecutter.project_slug}}.Application.Services;
+
+public class SubDomainEntityA : ISubDomainAService
+{
+    // Instantiate loggers and any required objects from the infrastructure and/or persistence layers, e.g.
+    // ```cs
+    // private readonly ILogger<SubDomainAService> _logger;
+    // private readonly SqlDbContext? _sqlDbContext;
+    // private readonly GcpBlobConnector? _gcpBlobConnetor;
+    // ```
+
+    // Set up the constructor, e.g.
+    // ```cs
+    // public SubDomainEntityA(SqlDbContext sqlDbContext)
+    // {
+    //     var loggerFactory = LoggerFactory.Create(builder => builder.AddConsole());
+    //     this._logger = loggerFactory.CreateLogger<SubDomainAService>();
+    //     try
+    //     {
+    //         this._sqlDbContext = sqlDbContext;
+    //         this._logger.LogInformation($"Initialization succeeded for {nameof(SubDomainAService)}");
+    //         // ...
+    //     }
+    //     catch (Exception ex)
+    //     {
+    //         this._logger.LogError(ex.Message);
+    //         this._sqlDbContext = null;
+    //     }
+    // }
+    // ```
+
+    /// <summary>
+    /// Creates a new SubdomainEntityA asynchronously.
+    /// </summary>
+    /// <param name="subDomainEntityA">The SubdomainEntityA to create.</param>
+    /// <returns>The created SubdomainEntityA, or null if creation fails.</returns>
+    public async Task<SubDomainEntityA?> CreateAsync(SubDomainEntityA? subDomainEntityA)
+    {
+        return await Task.FromResult<SubDomainEntityA?>(null);
+    }
+
+    /// <summary>
+    /// Gets SubDomainEntityA based on specified criteria asynchronously.
+    /// </summary>
+    /// <param name="filter">The criteria for filtering (optional).</param>
+    /// <returns>A collection of SubDomainEntityA matching the specified criteria, or null if no matches are found.</returns>
+    public async Task<IEnumerable<SubDomainEntityA>?> GetAsync(SubDomainAFilter? filter)
+    {
+        return await Task.FromResult<IEnumerable<SubDomainEntityA>?>(null);
+    }
+
+    /// <summary>
+    /// Gets SubDomainEntityA by its unique identifier asynchronously.
+    /// </summary>
+    /// <param name="id">The unique identifier of the SubdomainEntityA.</param>
+    /// <returns>The SubdomainEntityA with the specified identifier, or null if not found.</returns>
+    public async Task<SubDomainEntityA?> GetByIdAsync(Guid id)
+    {
+        return await Task.FromResult<SubDomainEntityA?>(null);
+    }
+
+    /// <summary>
+    /// Updates an existing SubDomainEntityA by its unique identifier asynchronously.
+    /// </summary>
+    /// <param name="id">The unique identifier of the SubdomainEntityA to update.</param>
+    /// <param name="subDomainEntityA">The new SubdomainEntityA.</param>
+    /// <returns>The updated SubdomainEntityA, or null if update fails.</returns>
+    public async Task<SubDomainEntityA?> UpdateByIdAsync(Guid id, SubDomainEntityA? subDomainEntityA)
+    {
+        return await Task.FromResult<SubDomainEntityA?>(null);
+    }
+
+    /// <summary>
+    /// Deletes SubDomainEntityA by its unique identifier asynchronously.
+    /// </summary>
+    /// <param name="id">The unique identifier of the SubdomainEntityA to delete.</param>
+    /// <returns>The deleted SubdomainEntityA, or null if deletion fails.</returns>
+    public async Task<SubDomainEntityA?> DeleteByIdAsync(Guid id)
+    {
+        return await Task.FromResult<SubDomainEntityA?>(null);
+    }
+}